move elasticity numerics files
