+++ SU2_CFD/include/numerics_structure.hpp
@@ -1351,40 +1351,10 @@ public:
   virtual void Compute_NodalStress_Term(CElement *element_container, CConfig *config);
 
   /*!
-   * \brief A virtual member to compute the plane stress term in an element for nonlinear structural problems
-   * \param[in] element_container - Element structure for the particular element integrated.
-   */
-  virtual void Compute_Plane_Stress_Term(CElement *element_container, CConfig *config);
-
-  /*!
-   * \brief A virtual member to compute the constitutive matrix in an element for structural problems
-   * \param[in] element_container - Element structure for the particular element integrated.
-   */
-  virtual void Compute_Constitutive_Matrix(CElement *element_container, CConfig *config);
-
-  /*!
-   * \brief A virtual member to compute the stress tensor in an element for structural problems
-   * \param[in] element_container - Element structure for the particular element integrated.
-   */
-  virtual void Compute_Stress_Tensor(CElement *element_container, CConfig *config);
-
-  /*!
-   * \brief A virtual member to compute the element-based Lame parameters and set the local properties
-   * \param[in] element_container - Element structure for the particular element integrated.
-   */
-  virtual void SetElement_Properties(CElement *element_container, CConfig *config);
-
-  /*!
    * \brief A virtual member to set the element-based local properties in mesh problems
    * \param[in] element_container - Element structure for the particular element integrated.
    */
   virtual void SetMeshElasticProperties(unsigned long iElem, su2double val_E);
-
-  /*!
-   * \brief A virtual member
-   * \param[in] config - Config structure
-   */
-  virtual void ReadDV(CConfig *config);
 
   /*!
    * \brief A virtual member to set the value of the design variables
@@ -1397,31 +1367,13 @@ public:
    * \brief A virtual member to retrieve the value of the design variables
    * \param[in] i_DV - Index of the design variable.
    */
-  virtual su2double Get_DV_Val(unsigned short i_DV);
-
-  /*!
-   * \brief A virtual member to add the Maxwell stress contribution
-   * \param[in] element_container - Element structure for the particular element integrated.
-   */
-  virtual void Add_MaxwellStress(CElement *element_container, CConfig *config);
-
-  /*!
-   * \brief A virtual member to set element-based electric field modulus
-   * \param[in] element_container - Element structure for the particular element integrated.
-   */
-  virtual void SetElectric_Properties(CElement *element_container, CConfig *config);
+  virtual su2double Get_DV_Val(unsigned short i_DV) const;
 
   /*!
    * \brief A virtual member to set the electric field
    * \param[in] EField_DV - New electric field computed by adjoint methods.
    */
   virtual void Set_ElectricField(unsigned short i_DV, su2double val_EField);
-
-  /*!
-   * \brief A virtual member to set the young modulus
-   * \param[in] val_Young - Value of the Young Modulus.
-   */
-  virtual void Set_YoungModulus(unsigned short i_DV, su2double val_Young);
 
   /*!
    * \brief A virtual member to set the material properties
@@ -4161,381 +4113,6 @@ public:
    */
   void ComputeResidual (su2double **val_stiffmatrix_elem, CConfig *config);
 };
-
-/*!
- * \class CFEAElasticity
- * \brief Generic class for computing the tangent matrix and the residual for structural problems
- * \ingroup FEM_Discr
- * \author R.Sanchez
- * \version 6.2.0 "Falcon"
- */
-class CFEAElasticity : public CNumerics {
-
-protected:
-
-  su2double E;              /*!< \brief Aux. variable, Young's modulus of elasticity. */
-  su2double Nu;             /*!< \brief Aux. variable, Poisson's ratio. */
-  su2double Rho_s;          /*!< \brief Aux. variable, Structural density. */
-  su2double Rho_s_DL;       /*!< \brief Aux. variable, Structural density (for dead loads). */
-  su2double Mu;             /*!< \brief Aux. variable, Lame's coeficient. */
-  su2double Lambda;         /*!< \brief Aux. variable, Lame's coeficient. */
-  su2double Kappa;          /*!< \brief Aux. variable, Compressibility constant. */
-
-  su2double *E_i;           /*!< \brief Young's modulus of elasticity. */
-  su2double *Nu_i;          /*!< \brief Poisson's ratio. */
-  su2double *Rho_s_i;       /*!< \brief Structural density. */
-  su2double *Rho_s_DL_i;    /*!< \brief Structural density (for dead loads). */
-
-  bool plane_stress;        /*!< \brief Checks if we are solving a plane stress case */
-
-  su2double **Ba_Mat,          /*!< \brief Matrix B for node a - Auxiliary. */
-  **Bb_Mat;                    /*!< \brief Matrix B for node b - Auxiliary. */
-  su2double *Ni_Vec;           /*!< \brief Vector of shape functions - Auxiliary. */
-  su2double **D_Mat;           /*!< \brief Constitutive matrix - Auxiliary. */
-  su2double **KAux_ab;         /*!< \brief Node ab stiffness matrix - Auxiliary. */
-  su2double **GradNi_Ref_Mat;  /*!< \brief Gradients of Ni - Auxiliary. */
-  su2double **GradNi_Curr_Mat; /*!< \brief Gradients of Ni - Auxiliary. */
-
-  su2double *FAux_Dead_Load;    /*!< \brief Auxiliar vector for the dead loads */
-
-  su2double *DV_Val;            /*!< \brief For optimization cases, value of the design variables. */
-  unsigned short n_DV;          /*!< \brief For optimization cases, number of design variables. */
-
-public:
-
-  /*!
-   * \brief Constructor of the class.
-   */
-  CFEAElasticity(void);
-
-  /*!
-   * \brief Constructor of the class (overload).
-   * \param[in] val_nDim - Number of dimensions of the problem.
-   * \param[in] val_nVar - Number of variables of the problem.
-   * \param[in] config - Definition of the particular problem.
-   */
-  CFEAElasticity(unsigned short val_nDim, unsigned short val_nVar, CConfig *config);
-
-  /*!
-   * \brief Destructor of the class.
-   */
-  virtual ~CFEAElasticity(void);
-
-  void SetMaterial_Properties(unsigned short iVal, su2double val_E, su2double val_Nu);
-
-  void SetMaterial_Density(unsigned short iVal, su2double val_Rho, su2double val_Rho_DL);
-
-  void Compute_Mass_Matrix(CElement *element_container, CConfig *config);
-
-  void Compute_Dead_Load(CElement *element_container, CConfig *config);
-
-  void Set_YoungModulus(unsigned short i_DV, su2double val_Young);
-
-  virtual void SetElement_Properties(CElement *element_container, CConfig *config);
-
-  void ReadDV(CConfig *config);
-
-  void Set_DV_Val(unsigned short i_DV, su2double val_DV);
-
-  su2double Get_DV_Val(unsigned short i_DV);
-
-  virtual void Compute_Tangent_Matrix(CElement *element_container, CConfig *config);
-
-  virtual void Compute_NodalStress_Term(CElement *element_container, CConfig *config);
-
-  virtual void Compute_Averaged_NodalStress(CElement *element_container, CConfig *config);
-
-  virtual void Compute_Plane_Stress_Term(CElement *element_container, CConfig *config);
-
-  virtual void Compute_Constitutive_Matrix(CElement *element_container, CConfig *config);
-  
-  virtual void Compute_Stress_Tensor(CElement *element_container, CConfig *config);
-
-	virtual void Add_MaxwellStress(CElement *element_container, CConfig *config);
-
-  virtual void SetElectric_Properties(CElement *element_container, CConfig *config);
-
-  virtual void Set_ElectricField(unsigned short i_DV, su2double val_EField);
-  
-protected:
-  void Compute_Lame_Parameters(void);
-
-};
-
-/*!
- * \class CFEALinearElasticity
- * \brief Class for computing the stiffness matrix of a linear, elastic problem.
- * \ingroup FEM_Discr
- * \author R.Sanchez
- * \version 6.2.0 "Falcon"
- */
-class CFEALinearElasticity : public CFEAElasticity {
-
-  su2double **nodalDisplacement;
-
-public:
-
-  /*!
-   * \brief Constructor of the class.
-   */
-  CFEALinearElasticity(void);
-
-  /*!
-   * \brief Constructor of the class (overload).
-   * \param[in] val_nDim - Number of dimensions of the problem.
-   * \param[in] val_nVar - Number of variables of the problem.
-   * \param[in] config - Definition of the particular problem.
-   */
-  CFEALinearElasticity(unsigned short val_nDim, unsigned short val_nVar, CConfig *config);
-
-  /*!
-   * \brief Destructor of the class.
-   */
-  ~CFEALinearElasticity(void);
-
-  void Compute_Tangent_Matrix(CElement *element_container, CConfig *config);
-
-  void Compute_Constitutive_Matrix(CElement *element_container, CConfig *config);
-
-  void Compute_Averaged_NodalStress(CElement *element_container, CConfig *config);
-
-};
-
-/*!
- * \class CFEANonlinearElasticity
- * \brief Class for computing the stiffness matrix of a nonlinear, elastic problem.
- * \ingroup FEM_Discr
- * \author R.Sanchez
- * \version 6.2.0 "Falcon"
- */
-class CFEANonlinearElasticity : public CFEAElasticity {
-
-protected:
-
-  su2double **F_Mat;             /*!< \brief Deformation gradient. */
-  su2double **b_Mat;             /*!< \brief Left Cauchy-Green Tensor. */
-  su2double **currentCoord;      /*!< \brief Current coordinates. */
-  su2double **Stress_Tensor;     /*!< \brief Cauchy stress tensor */
-
-  su2double **FmT_Mat;           /*!< \brief Deformation gradient inverse and transpose. */
-
-  su2double **KAux_P_ab;         /*!< \brief Auxiliar matrix for the pressure term */
-  su2double *KAux_t_a;           /*!< \brief Auxiliar matrix for the pressure term */
-
-  su2double J_F;                 /*!< \brief Jacobian of the transformation (determinant of F) */
-
-  su2double f33;                 /*!< \brief Plane stress term for non-linear 2D plane stress analysis */
-
-  bool nearly_incompressible;    /*!< \brief Boolean to consider nearly_incompressible effects */
-
-  su2double **F_Mat_Iso;         /*!< \brief Isocoric component of the deformation gradient. */
-  su2double **b_Mat_Iso;         /*!< \brief Isocoric component of the left Cauchy-Green tensor. */
-
-  su2double C10, D1;             /*!< \brief C10 = Mu/2. D1 = Kappa/2. */
-  su2double J_F_Iso;             /*!< \brief J_F_Iso: det(F)^-1/3. */
-
-  su2double ****cijkl;           /*!< \brief Constitutive tensor i,j,k,l (defined only for incompressibility - near inc.). */
-
-  bool maxwell_stress;           /*!< \brief Consider the effects of the dielectric loads */
-
-  su2double *EField_Ref_Unit,    /*!< \brief Electric Field, unitary, in the reference configuration. */
-  *EField_Ref_Mod;               /*!< \brief Electric Field, modulus, in the reference configuration. */
-  su2double *EField_Curr_Unit;   /*!< \brief Auxiliary vector for the unitary Electric Field in the current configuration. */
-  unsigned short nElectric_Field,
-  nDim_Electric_Field;
-
-  su2double *ke_DE_i;           /*!< \brief Electric Constant for Dielectric Elastomers. */
-
-  su2double ke_DE;              /*!< \brief Electric Constant for Dielectric Elastomers. */
-  su2double EFieldMod_Ref;      /*!< \brief Modulus of the electric field in the reference configuration. */
-
-
-public:
-
-  /*!
-   * \brief Constructor of the class.
-   * \param[in] val_nDim - Number of dimensions of the problem.
-   * \param[in] val_nVar - Number of variables of the problem.
-   * \param[in] config - Definition of the particular problem.
-   */
-  CFEANonlinearElasticity(unsigned short val_nDim, unsigned short val_nVar, CConfig *config);
-
-  /*!
-   * \brief Destructor of the class.
-   */
-  virtual ~CFEANonlinearElasticity(void);
-
-  void Compute_Tangent_Matrix(CElement *element_container, CConfig *config);
-
-  void Compute_NodalStress_Term(CElement *element_container, CConfig *config);
-
-  void Compute_Averaged_NodalStress(CElement *element_container, CConfig *config);
-
-  void Add_MaxwellStress(CElement *element_container, CConfig *config);
-
-  void SetElectric_Properties(CElement *element_container, CConfig *config);
-
-  void Compute_FmT_Mat(void);
-
-  void Compute_Isochoric_F_b(void);
-
-  void Assign_cijkl_D_Mat(void);
-
-  void Set_ElectricField(unsigned short i_DV, su2double val_EField);
-
-  void Set_YoungModulus(unsigned short i_DV, su2double val_Young);
-
-  void SetMaterial_Properties(unsigned short iVal, su2double val_E, su2double val_Nu);
-
-  void SetMaterial_Density(unsigned short iVal, su2double val_Rho, su2double val_Rho_DL);
-
-  su2double deltaij(unsigned short iVar, unsigned short jVar);
-
-  virtual void Compute_Plane_Stress_Term(CElement *element_container, CConfig *config);
-
-  virtual void Compute_Constitutive_Matrix(CElement *element_container, CConfig *config);
-
-  virtual void Compute_Stress_Tensor(CElement *element_container, CConfig *config);
-
-
-};
-
-/*!
- * \class CFEM_NeoHookean_Comp
- * \brief Class for computing the constitutive and stress tensors for a neo-Hookean material model, compressible.
- * \ingroup FEM_Discr
- * \author R.Sanchez
- * \version 6.2.0 "Falcon"
- */
-class CFEM_NeoHookean_Comp : public CFEANonlinearElasticity {
-
-public:
-
-  /*!
-   * \brief Constructor of the class.
-   * \param[in] val_nDim - Number of dimensions of the problem.
-   * \param[in] val_nVar - Number of variables of the problem.
-   * \param[in] config - Definition of the particular problem.
-   */
-  CFEM_NeoHookean_Comp(unsigned short val_nDim, unsigned short val_nVar, CConfig *config);
-
-  /*!
-   * \brief Destructor of the class.
-   */
-  ~CFEM_NeoHookean_Comp(void);
-
-  void Compute_Plane_Stress_Term(CElement *element_container, CConfig *config);
-
-  void Compute_Constitutive_Matrix(CElement *element_container, CConfig *config);
-  using CNumerics::Compute_Constitutive_Matrix;
-
-  void Compute_Stress_Tensor(CElement *element_container, CConfig *config);
-
-};
-
-/*!
- * \class CFEM_IdealDE
- * \brief Class for computing the constitutive and stress tensors for a nearly-incompressible ideal DE.
- * \ingroup FEM_Discr
- * \author R.Sanchez
- * \version 6.2.0 "Falcon"
- */
-class CFEM_IdealDE : public CFEANonlinearElasticity {
-
-	su2double trbbar, Eg, Eg23, Ek, Pr;	/*!< \brief Variables of the model calculation. */
-
-public:
-
-  /*!
-   * \brief Constructor of the class.
-   * \param[in] val_nDim - Number of dimensions of the problem.
-   * \param[in] val_nVar - Number of variables of the problem.
-   * \param[in] config - Definition of the particular problem.
-   */
-  CFEM_IdealDE(unsigned short val_nDim, unsigned short val_nVar, CConfig *config);
-
-  /*!
-   * \brief Destructor of the class.
-   */
-  ~CFEM_IdealDE(void);
-
-  void Compute_Plane_Stress_Term(CElement *element_container, CConfig *config);
-
-  void Compute_Constitutive_Matrix(CElement *element_container, CConfig *config);
-
-  void Compute_Stress_Tensor(CElement *element_container, CConfig *config);
-
-};
-
-/*!
- * \class CFEM_NeoHookean_Comp
- * \brief Class for computing the constitutive and stress tensors for a Knowles stored-energy function, nearly incompressible.
- * \ingroup FEM_Discr
- * \author R.Sanchez
- * \version 6.2.0 "Falcon"
- */
-class CFEM_Knowles_NearInc : public CFEANonlinearElasticity {
-
-	su2double trbbar, term1, term2, Ek, Pr;	/*!< \brief Variables of the model calculation. */
-	su2double Bk, Nk;						/*!< \brief Parameters b and n of the model. */
-
-public:
-
-  /*!
-   * \brief Constructor of the class.
-   * \param[in] val_nDim - Number of dimensions of the problem.
-   * \param[in] val_nVar - Number of variables of the problem.
-   * \param[in] config - Definition of the particular problem.
-   */
-  CFEM_Knowles_NearInc(unsigned short val_nDim, unsigned short val_nVar, CConfig *config);
-
-  /*!
-   * \brief Destructor of the class.
-   */
-  ~CFEM_Knowles_NearInc(void);
-
-  void Compute_Plane_Stress_Term(CElement *element_container, CConfig *config);
-
-  void Compute_Constitutive_Matrix(CElement *element_container, CConfig *config);
-  using CNumerics::Compute_Constitutive_Matrix;
-
-	void Compute_Stress_Tensor(CElement *element_container, CConfig *config);
-
-};
-
-/*!
- * \class CFEM_DielectricElastomer
- * \brief Class for computing the constitutive and stress tensors for a dielectric elastomer.
- * \ingroup FEM_Discr
- * \author R.Sanchez
- * \version 6.2.0 "Falcon"
- */
-class CFEM_DielectricElastomer : public CFEANonlinearElasticity {
-
-public:
-
-  /*!
-   * \brief Constructor of the class.
-   * \param[in] val_nDim - Number of dimensions of the problem.
-   * \param[in] val_nVar - Number of variables of the problem.
-   * \param[in] config - Definition of the particular problem.
-   */
-  CFEM_DielectricElastomer(unsigned short val_nDim, unsigned short val_nVar, CConfig *config);
-
-  /*!
-   * \brief Destructor of the class.
-   */
-  ~CFEM_DielectricElastomer(void);
-
-  void Compute_Plane_Stress_Term(CElement *element_container, CConfig *config);
-
-  void Compute_Constitutive_Matrix(CElement *element_container, CConfig *config);
-  using CNumerics::Compute_Constitutive_Matrix;
-
-  void Compute_Stress_Tensor(CElement *element_container, CConfig *config);
-
-};
-
 
 /*!
  * \class CSourceNothing
