+++ SU2_CFD/src/meson.build
@@ -26,9 +26,6 @@ su2_cfd_src = files(['solver_direct_transition.cpp',
                      'solver_direct_heat.cpp',
                      'numerics_direct_heat.cpp',
                      'fluid_model_ppr.cpp',
-                     'numerics_direct_elasticity.cpp',
-                     'numerics_direct_turbulent.cpp',
-                     'numerics_direct_elasticity_linear.cpp',
                      'python_wrapper_structure.cpp',
                      'numerics_structure.cpp',
                      'numerics_template.cpp',
@@ -52,49 +49,56 @@ su2_cfd_src += files(['output/CAdjElasticityOutput.cpp',
                       'output/output_physics.cpp',
                       'output/COutput.cpp',
                       'output/output_structure_legacy.cpp',
-		      'output/filewriter/CParallelDataSorter.cpp',
-		      'output/filewriter/CFVMDataSorter.cpp',
-		      'output/filewriter/CFEMDataSorter.cpp',
-		      'output/filewriter/CSurfaceFEMDataSorter.cpp',
-		      'output/filewriter/CSurfaceFVMDataSorter.cpp',
-		      'output/filewriter/CParallelFileWriter.cpp',
-		      'output/filewriter/CParaviewFileWriter.cpp',
-		      'output/filewriter/CParaviewBinaryFileWriter.cpp',
-		      'output/filewriter/CTecplotFileWriter.cpp',
-		      'output/filewriter/CTecplotBinaryFileWriter.cpp',
-		      'output/filewriter/CCSVFileWriter.cpp',
-		      'output/filewriter/CSU2FileWriter.cpp',
-		      'output/filewriter/CSU2BinaryFileWriter.cpp',
-		      'output/filewriter/CSU2MeshFileWriter.cpp'])
+                      'output/filewriter/CParallelDataSorter.cpp',
+                      'output/filewriter/CFVMDataSorter.cpp',
+                      'output/filewriter/CFEMDataSorter.cpp',
+                      'output/filewriter/CSurfaceFEMDataSorter.cpp',
+                      'output/filewriter/CSurfaceFVMDataSorter.cpp',
+                      'output/filewriter/CParallelFileWriter.cpp',
+                      'output/filewriter/CParaviewFileWriter.cpp',
+                      'output/filewriter/CParaviewBinaryFileWriter.cpp',
+                      'output/filewriter/CTecplotFileWriter.cpp',
+                      'output/filewriter/CTecplotBinaryFileWriter.cpp',
+                      'output/filewriter/CCSVFileWriter.cpp',
+                      'output/filewriter/CSU2FileWriter.cpp',
+                      'output/filewriter/CSU2BinaryFileWriter.cpp',
+                      'output/filewriter/CSU2MeshFileWriter.cpp'])
 
 su2_cfd_src += files(['variables/CIncNSVariable.cpp',
-                     'variables/CTransLMVariable.cpp',
-	                   'variables/CAdjEulerVariable.cpp',
-                     'variables/CHeatFVMVariable.cpp',
-                     'variables/CTurbVariable.cpp',
-                     'variables/CAdjNSVariable.cpp',
-                     'variables/CBaselineVariable.cpp',
-                     'variables/CDiscAdjFEAVariable.cpp',
-                     'variables/CDiscAdjFEABoundVariable.cpp',
-                     'variables/CDiscAdjMeshBoundVariable.cpp',
-                     'variables/CFEABoundVariable.cpp',
-                     'variables/CDiscAdjVariable.cpp',
-                     'variables/CTurbSAVariable.cpp',
-                     'variables/CFEAVariable.cpp',
-                     'variables/CMeshVariable.cpp',
-                     'variables/CMeshBoundVariable.cpp',
-                     'variables/CMeshElement.cpp',
-                     'variables/CTurbSSTVariable.cpp',
-                     'variables/CVariable.cpp',
-                     'variables/CNSVariable.cpp',
-                     'variables/CAdjTurbVariable.cpp',
-                     'variables/CIncEulerVariable.cpp',
-                     'variables/CEulerVariable.cpp'])
+                      'variables/CTransLMVariable.cpp',
+	                    'variables/CAdjEulerVariable.cpp',
+                      'variables/CHeatFVMVariable.cpp',
+                      'variables/CTurbVariable.cpp',
+                      'variables/CAdjNSVariable.cpp',
+                      'variables/CBaselineVariable.cpp',
+                      'variables/CDiscAdjFEAVariable.cpp',
+                      'variables/CDiscAdjFEABoundVariable.cpp',
+                      'variables/CDiscAdjMeshBoundVariable.cpp',
+                      'variables/CFEABoundVariable.cpp',
+                      'variables/CDiscAdjVariable.cpp',
+                      'variables/CTurbSAVariable.cpp',
+                      'variables/CFEAVariable.cpp',
+                      'variables/CMeshVariable.cpp',
+                      'variables/CMeshBoundVariable.cpp',
+                      'variables/CMeshElement.cpp',
+                      'variables/CTurbSSTVariable.cpp',
+                      'variables/CVariable.cpp',
+                      'variables/CNSVariable.cpp',
+                      'variables/CAdjTurbVariable.cpp',
+                      'variables/CIncEulerVariable.cpp',
+                      'variables/CEulerVariable.cpp'])
 
 su2_cfd_src += files(['solvers/CMeshSolver.cpp',
-                     'solvers/CDiscAdjMeshSolver.cpp'])
+                      'solvers/CDiscAdjMeshSolver.cpp'])
 
-su2_cfd_src += files(['numerics/CFEAMeshElasticity.cpp'])
+su2_cfd_src += files(['numerics/elasticity/CFEAElasticity.cpp',
+                      'numerics/elasticity/CFEALinearElasticity.cpp',
+                      'numerics/elasticity/CFEAMeshElasticity.cpp',
+                      'numerics/elasticity/CFEANonlinearElasticity.cpp',
+                      'numerics/elasticity/CFEM_NeoHookean_Comp.cpp',
+                      'numerics/elasticity/CFEM_Knowles_NearInc.cpp',
+                      'numerics/elasticity/CFEM_IdealDE.cpp',
+                      'numerics/elasticity/CFEM_DielectricElastomer.cpp'])
 
 su2_cfd_src += files(['interfaces/CInterface.cpp',
                       'interfaces/cfd/CConservativeVarsInterface.cpp',
@@ -117,23 +121,24 @@ su2_cfd_src += files(['drivers/CDriver.cpp',
 if get_option('enable-normal')
   su2_cfd = executable('SU2_CFD',
                        su2_cfd_src, 
-      		       install : true, 
-		       dependencies : [su2_deps, common_dep], 
-		       cpp_args:  ['-fPIC'] + [default_warning_flags, su2_cpp_args])
+      		             install : true, 
+              	       dependencies : [su2_deps, common_dep], 
+		                   cpp_args:  ['-fPIC'] + [default_warning_flags, su2_cpp_args])
 endif
 
 if get_option('enable-autodiff')
   su2_cfd_ad = executable('SU2_CFD_AD',
-                          su2_cfd_src,
-			  install : true, 
-			  dependencies : [su2_deps, codi_dep, commonAD_dep], 
-			  cpp_args: ['-fPIC'] + [default_warning_flags, su2_cpp_args, codi_rev_args])
+                       su2_cfd_src,
+                       install : true, 
+                       dependencies : [su2_deps, codi_dep, commonAD_dep], 
+                       cpp_args: ['-fPIC'] + [default_warning_flags, su2_cpp_args, codi_rev_args])
 endif
 
 if get_option('enable-directdiff')
   su2_cfd_dd = executable('SU2_CFD_DIRECTDIFF',
-                          su2_cfd_src,
-			  install : true, 
-			  dependencies : [su2_deps, codi_dep, commonDD_dep], 
-			  cpp_args: ['-fPIC'] + [default_warning_flags, su2_cpp_args, codi_for_args])
+                       su2_cfd_src,
+                       install : true, 
+                       dependencies : [su2_deps, codi_dep, commonDD_dep], 
+                       cpp_args: ['-fPIC'] + [default_warning_flags, su2_cpp_args, codi_for_args])
 endif
+
